01.25.01

+++ Defence/Assets/Defence/Scripts/Enemy/Enemy.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 public class Enemy : MonoBehaviour
 {
     public static Enemy instance;
-    public int Enemy_HP;
+    public float Enemy_HP;
     public int Enemy_att;
     public Slider _hpbar;
 
@@ -18,7 +18,7 @@ public class Enemy : MonoBehaviour
     void Start()
     {
         instance = this;
-        Enemy_HP = 200+(info.instance._round*50);
+        Enemy_HP = 200.0f+(info.instance._round*50.0f);
         Enemy_att = Random.Range(10 * EnemyManager.instance._round, 20 * EnemyManager.instance._round);
         _hpbar = this.transform.GetChild(1).transform.GetChild(0).GetComponent<Slider>();
         _hpbar.maxValue = Enemy_HP;
